backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pydantic
pydantic-settings
kafka-python
pandas
scikit-learn
joblib
shap
python-dotenv
